% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all scripts}
\usage{
run_all(dir = "R.d", verbose = TRUE)
}
\arguments{
\item{dir}{Directory to use for R.d files (default = R.d)}

\item{verbose}{Should the files processed be printed.}
}
\value{
Logical value (TRUE) invisibly
}
\description{
Run all scripts
}
\examples{
\dontrun{
run_all(dir="~/mydir/R.dx")
}
}
